pyproject.toml
[tool.poetry]
name = "kubechat"
version = "0.1.0"
description = "A LLM based tool to support Text2SQL, Chatbot, KubeBlocks pilot and troubleshooting"
authors = ["nayutah <dullboy@apecloud.com>"]
readme = "README.md"
packages = [{ include = "server" }]

[tool.poetry.dependencies]
python = ">=3.10.0,<4.0"
langchain = "^0.0.271"
llama-index = "^0.6.35"
uvicorn = {extras = ["standard"], version = "^0.23.1"}
django = "^4.2.2"
pydantic = "^1.10.9"
sqlalchemy = "^2.0.16"
sqlparse = "^0.4.4"
llama-hub = "^0.0.3"
pymysql = "^1.0.3"
cryptography = "^41.0.1"
python-dotenv = "^1.0.0"
django-environ = "^0.10.0"
django-ninja = "^0.22.2"
requests = "^2.31.0"
auth0-python = "^4.2.0"
celery = "^5.3.1"
channels = "^4.0.0"
pypdf = "^3.11.0"
sentence-transformers = "^2.2.2"
qdrant-client = {version="^1.3.0", python = ">=3.8.1,<3.12"}
clickhouse-connect = { version = "^0.5.14", optional = true }
pymilvus = { version = "^2.2.12", optional = true }
redis = "^4.6.0"
pymongo = { version = "^4.4.0", optional = true }
elasticsearch = { version = "^8.8.0", optional = true }
weaviate-client = { version = "^3.21.0", optional = true }
arrow = "^1.2.3"
func-timeout = "^4.3.5"
docx2txt = "^0.8"
terminal = "^0.4.0"
types-chardet = "^5.0.4.6"
pysocks = "^1.7.1"
psycopg2-binary = "^2.9.6"
watchfiles = "^0.19.0"
flower = "^2.0.0"
django-celery-beat = "^2.5.0"
boto3 = "^1.26.165"
oss2 = "^2.18.0"
mypy = "^1.4.1"
django-cors-headers = "^4.1.0"
socksio = "^1.0.0"
transformers-stream-generator = "^0.0.4"
django-simpleui = "^2023.3.1"
django-prometheus = "^2.3.1"
whitenoise = "^6.5.0"
types-requests = "^2.31.0.2"
lark-oapi = "^1.0.11"
text2vec = "^1.2.2"
pyarmor = "^8.3.1"
pytablewriter = "^1.0.0"
tabulate = "^0.9.0"
gradio = "^3.41.2"
python-docx = "^0.8.11"

[tool.poetry.group.llmserver]
optional = true

[tool.poetry.group.llmserver.dependencies]
fastapi = "^0.97.0"
nltk = "^3.8.1"
torch = ">=1,<3"
accelerate = "^0.20.3"
gpt4all = "0.3.5"
auto-gpt-plugin-template = "^0.0.3"
transformers = "^4.30.2"

[tool.poetry.group.lint.dependencies]
ruff = "^0.0.276"
types-toml = "^0.10.8.1"
types-redis = "^4.3.21.6"
black = "^23.1.0"
types-chardet = "^5.0.4.6"

[tool.poetry.extras]
all = [
    "pymilvus",
    "pymongo",
    "chromadb",
    "elasticsearch",
    "weaviate-client",
    "clickhouse-connect"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "server.main:start"
llmserver = "server.llmserver:start"
webserver = "server.webserver:start"
test_local_path_embedding = "readers.test_local_path_embedding:start"

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"


[[tool.poetry.source]]
name = "PyPI"
priority = "supplemental"

[tool.ruff]
select = [
    "E",  # pycodestyle
    "F",  # pyflakes
    "I",  # isort
]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 120

[tool.mypy]
ignore_missing_imports = "True"
disallow_untyped_defs = "True"
exclude = ["playground"]
